requirements.txt
cffi==1.17.1
comtypes==1.4.8
numpy==2.1.3
pillow==11.0.0
pycparser==2.22
PyGetWindow==0.0.9
PyRect==0.2.0
pystray==0.19.5
six==1.16.0
SoundCard==0.4.3
